+++ lib/lightning/scrubber.ex
@@ -16,22 +16,27 @@ defmodule Lightning.Scrubber do
   """
   defmodule State do
     @moduledoc false
-    @type t :: {
-            samples :: [String.t()]
-          }
+    @typep samples :: [String.t()]
+    @type t :: {samples()}
 
     @spec new(samples :: [String.t()]) :: t()
     def new(samples) do
       {samples}
     end
 
-    @spec samples(state :: t()) :: [String.t()]
-    def samples({samples}), do: samples
+    @spec add_samples(state :: t(), samples()) :: t()
+    def add_samples({samples}, new_samples) do
+      [samples, new_samples]
+      |> Enum.concat()
+      |> Enum.uniq()
+      |> Enum.sort_by(&String.length/1, :desc)
+      |> then(fn samples -> {samples} end)
+    end
 
-    @spec scrub(data :: String.t(), state :: t()) :: String.t()
-    def scrub(nil, _state), do: nil
+    @spec scrub(state :: t(), data :: String.t() | nil) :: String.t()
+    def scrub(_state, nil), do: nil
 
-    def scrub(string, {samples}) do
+    def scrub({samples}, string) when is_binary(string) do
       samples
       |> Enum.reduce(string, fn x, acc ->
         String.replace(acc, x, "***", global: true)
@@ -41,10 +46,17 @@ defmodule Lightning.Scrubber do
 
   use Agent
 
-  @spec start_link(opts :: [samples: [String.t()], name: nil | GenServer.name()]) ::
+  @spec start_link(
+          opts :: [
+            samples: [String.t()],
+            basic_auth: [String.t()],
+            name: nil | GenServer.name()
+          ]
+        ) ::
           Agent.on_start()
   def start_link(opts) do
     samples = Keyword.get(opts, :samples, [])
+    basic_auth = Keyword.get(opts, :basic_auth, [])
 
     server_opts =
       Keyword.get(opts, :name)
@@ -57,30 +69,37 @@ defmodule Lightning.Scrubber do
       end
 
     Agent.start_link(
-      fn -> State.new(samples |> encode_samples()) end,
+      fn -> State.new(samples |> encode_samples(basic_auth)) end,
       server_opts
     )
   end
 
   def samples(agent) do
-    Agent.get(agent, &State.samples/1)
+    Agent.get(agent, & &1) |> elem(0)
+  end
+
+  def add_samples(agent, new_samples, basic_auth) do
+    new_encoded_samples = encode_samples(new_samples, basic_auth)
+    Agent.update(agent, &State.add_samples(&1, new_encoded_samples))
   end
 
   def scrub(agent, lines) when is_list(lines) do
-    samples = samples(agent)
-    lines |> Enum.map(fn line -> State.scrub(line, {samples}) end)
+    state = Agent.get(agent, & &1)
+    lines |> Enum.map(fn line -> State.scrub(state, line) end)
   end
 
   def scrub(agent, data) do
-    agent |> Agent.get(&State.scrub(data, &1))
+    agent |> Agent.get(&State.scrub(&1, data))
   end
 
   @doc """
   Prepare a list of sensitive samples (strings) into a potentially bigger list
   composed of variations a sample may appear.
   """
-  @spec encode_samples(samples :: [String.t()]) :: [String.t()]
-  def encode_samples(samples) do
+  @spec encode_samples(samples :: [String.t()], basic_auth :: [String.t()]) :: [
+          String.t()
+        ]
+  def encode_samples(samples, basic_auth \\ []) do
     stringified_samples =
       samples
       |> Enum.filter(fn x -> not is_boolean(x) end)
@@ -93,6 +112,11 @@ defmodule Lightning.Scrubber do
         "#{x}:#{y}"
         |> Base.encode64()
       end)
+      |> Enum.concat(basic_auth)
+      |> Enum.concat(
+        stringified_samples
+        |> Enum.map(fn x -> Base.encode64(x) end)
+      )
 
     Enum.concat([stringified_samples, base64_secrets])
     |> Enum.sort_by(&String.length/1, :desc)